requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
fastapi==0.95.0
Flask==2.2.3
frozenlist==1.3.3
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.10
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
modal-client==0.48.1679
multidict==6.0.4
protobuf==4.22.1
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.25.1
rich==13.3.2
sniffio==1.3.0
starlette==0.26.1
synchronicity==0.3.1
tblib==1.7.0
toml==0.10.2
twilio==7.17.0
typeguard==3.0.2
typer==0.7.0
types-certifi==2021.10.8.3
types-toml==0.10.8.5
typing_extensions==4.5.0
urllib3==1.26.15
watchfiles==0.18.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
